# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EI-Nexus API
  description: >
    Autonomous Engineering Intelligence Twin — REST API for fleet telemetry,
    fault simulation, the 4-tool MCP root-cause agent, the autonomous
    closed-loop agent, ARIA (chat assistant), evaluation metrics, ROI, and
    the SQLite-backed persistence layer.
  version: "1.3.0"
  contact:
    name: Team MCP Mavericks
servers:
  - url: http://localhost:8000
    description: Local development server (python run.py)
...
paths:
...
  /api/telemetry:
    get:
      tags: [Fleet & Telemetry]
      summary: Full telemetry + history for one asset
      parameters:
        - name: asset_id
          in: query
          required: false
          schema:
            type: string
          description: Defaults to the first fleet asset if omitted
      responses:
        "200":
          description: Telemetry snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelemetrySnapshot"
  /api/inject_fault:
    post:
      tags: [Fleet & Telemetry]
      summary: Simulate a field failure on a specific asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                asset_id: { type: string, example: "MC-2201-001" }
                failure_mode:
                  type: string
                  enum: [OVERTEMP, VIBRATION, VOLTAGE, CURRENT]
      responses:
        "200":
          description: Fault injected
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  failure_mode: { type: string }
                  asset_id: { type: string }
  /api/apply_fix:
    post:
      tags: [Fleet & Telemetry]
      summary: Manually apply the fix on a specific asset
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                asset_id: { type: string }
      responses:
        "200":
          description: Fix applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  asset_id: { type: string }
...
  /api/analyze_rca:
    post:
      tags: [RCA Agent]
      summary: Run the 4-tool MCP agent once against a symptom query
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: Free-text field symptom. If omitted, auto-built from the live twin's current anomaly.
                use_llm:
                  type: boolean
                  default: true
                  description: Attempt Gemini synthesis; falls back to rule-based if unavailable.
      responses:
        "200":
          description: RCA result with live reasoning trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RCAResult"
...
  /api/autonomous/start:
    post:
      tags: [Autonomous Mode]
      summary: Engage the closed-loop autonomous agent
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                auto_inject:
                  type: boolean
                  default: true
                  description: Auto-simulate a fault if an asset stays nominal too long
      responses:
        "200":
          description: Engaged
          content:
            application/json:
              schema: { type: object, properties: { ok: { type: boolean } } }
...
  /api/chat:
    post:
      tags: [ARIA Chatbot]
      summary: Ask ARIA a question about the fleet or incident history
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message: { type: string, example: "what happened?" }
                use_llm: { type: boolean, default: true }
      responses:
        "200":
          description: ARIA's reply
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply: { type: string }
                  used_llm: { type: boolean }
                  bot_name: { type: string, example: "ARIA" }
...
components:
  schemas:
...
    TelemetrySnapshot:
      type: object
      properties:
        asset_id: { type: string }
        site: { type: string }
        state:
          type: object
          properties:
            temperature: { type: number }
            vibration: { type: number }
            load: { type: number }
            voltage: { type: number }
            current: { type: number }
            fan_speed: { type: number }
        status: { type: string }
        active_fault: { type: string, nullable: true }
        history: { type: array, items: { type: object } }
        anomalies: { type: array, items: { type: object } }
    RCAResult:
      type: object
      properties:
        trace: { type: array, items: { type: object }, description: "Step-by-step tool-call trace" }
        inferred_failure_mode: { type: string }
        mode_confidence: { type: number }
        synthesis: { type: string }
        llm_used: { type: boolean }
        citations: { type: array, items: { type: string } }
        total_latency_ms: { type: number }
...